Chore/document update (#4)
* wip

* upgrade deps and optim images

* readme updated
