--- client/src/entry/css/App.scss
@@ -1,4 +1,0 @@
-html,
-body {
-  margin: 0;
-}
