--- client/src/entry/js/components/ClientRouting.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
-import { renderRoutes } from 'react-router-config';
-
-import routes from './routes';
-
-const ClientRouting = () => (
-  <BrowserRouter>
-    {renderRoutes(routes)}
-  </BrowserRouter>
-);
-
-export default ClientRouting;
