--- client/src/entry/js/polyfills.js
@@ -1,1 +1,0 @@
-import "@babel/polyfill";
