--- client/src/views/home/css/Home.scss
@@ -1,20 +1,0 @@
-.cover {
-  color: white;
-  background: no-repeat center bottom;
-  background-size: cover;
-  background-image: url('../img/cover.jpg');
-  height: 100vh;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  > div {
-    margin-bottom: 10px;
-  }
-
-  .logo {
-    height: 100px;
-  }
-}
