--- client/src/views/page1/css/Page1.scss
@@ -1,16 +1,0 @@
-.cover {
-  color: white;
-  background: no-repeat center bottom;
-  background-size: cover;
-  background-image: url('../img/cover.jpg');
-  height: 100vh;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  > div {
-    margin-bottom: 10px;
-  }
-}
