+++ package.json
@@ -12,51 +12,51 @@
     "body-parser": "1.18.3",
     "compression": "1.7.3",
     "consolidate": "0.15.1",
-    "express": "4.16.3",
+    "express": "4.16.4",
     "handlebars": "4.0.12",
     "prop-types": "15.6.2",
-    "react": "16.5.2",
-    "react-dom": "16.5.2",
+    "react": "16.6.3",
+    "react-dom": "16.6.3",
     "react-helmet": "5.2.0",
     "react-router-config": "1.0.0-beta.4",
     "react-router-dom": "4.3.1"
   },
   "devDependencies": {
-    "@babel/core": "7.1.0",
+    "@babel/core": "7.1.6",
     "@babel/plugin-proposal-class-properties": "7.1.0",
     "@babel/plugin-syntax-dynamic-import": "7.0.0",
     "@babel/plugin-transform-react-jsx-source": "7.0.0",
-    "@babel/preset-env": "7.1.0",
+    "@babel/preset-env": "7.1.6",
     "@babel/preset-react": "7.0.0",
     "@babel/preset-typescript": "7.1.0",
-    "autoprefixer": "9.1.5",
+    "autoprefixer": "9.3.1",
     "babel-core": "7.0.0-bridge.0",
-    "babel-loader": "8.0.2",
+    "babel-loader": "8.0.4",
     "circular-dependency-plugin": "5.0.2",
     "cross-env": "5.2.0",
-    "css-loader": "1.0.0",
+    "css-loader": "1.0.1",
     "fast-stable-stringify": "1.0.0",
     "file-loader": "2.0.0",
     "image-webpack-loader": "3.4.2",
-    "mini-css-extract-plugin": "0.4.3",
-    "node-sass": "4.9.3",
+    "mini-css-extract-plugin": "0.4.4",
+    "node-sass": "4.10.0",
     "optimize-css-assets-webpack-plugin": "5.0.1",
     "postcss-flexbugs-fixes": "4.1.0",
     "postcss-loader": "3.0.0",
     "sass-loader": "7.1.0",
     "terser-webpack-plugin": "1.1.0",
-    "typescript": "3.0.3",
-    "url-loader": "1.1.1",
-    "webpack": "4.20.0",
-    "webpack-bundle-analyzer": "3.0.2",
-    "webpack-cli": "3.1.1",
+    "typescript": "3.1.6",
+    "url-loader": "1.1.2",
+    "webpack": "4.26.0",
+    "webpack-bundle-analyzer": "3.0.3",
+    "webpack-cli": "3.1.2",
     "webpack-dev-middleware": "3.4.0",
-    "webpack-hot-middleware": "2.24.2",
+    "webpack-hot-middleware": "2.24.3",
     "webpack-hot-server-middleware": "0.5.0",
     "webpack-manifest-plugin": "2.0.4",
     "webpack-merge": "4.1.4",
     "webpack-node-externals": "1.7.2",
-    "webpack-notifier": "1.6.0",
+    "webpack-notifier": "1.7.0",
     "webpackbar": "2.6.3",
     "xxhashjs": "0.2.2"
   },
