+++ src/entry/css/App.scss
@@ -1,0 +1,4 @@
+html,
+body {
+  margin: 0;
+}
