+++ src/entry/js/components/Bundles.js
@@ -1,0 +1,6 @@
+import syncComponent from 'src/hoc/code-splitting/js/syncComponent';
+
+export const MainLayout = syncComponent('MainLayout', require('src/views/main-layout/js/MainLayout'));
+export const Home = syncComponent('Home', require('src/views/home/js/Home'));
+export const Page1 = syncComponent('Page1', require('src/views/page1/js/Page1'));
+export const Page2 = syncComponent('Page2', require('src/views/page2/js/Page2'));
