+++ src/entry/js/polyfills.js
@@ -1,0 +1,1 @@
+import "@babel/polyfill";
