+++ src/hoc/code-splitting/js/syncComponent.js
@@ -1,0 +1,26 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+
+const syncComponent = (chunkName, mod) => {
+  const Component = mod.default ? mod.default : mod;
+
+  const SyncComponent = ({ staticContext, ...otherProps }) => {
+    if (staticContext.splitPoints) {
+      staticContext.splitPoints.push(chunkName);
+    }
+
+    return <Component {...otherProps} />;
+  };
+
+  SyncComponent.propTypes = {
+    staticContext: PropTypes.object,
+  };
+
+  SyncComponent.defaultProps = {
+    staticContext: undefined,
+  };
+
+  return SyncComponent;
+};
+
+export default syncComponent;
