+++ src/utils/EnvUtils.js
@@ -1,0 +1,13 @@
+let serverSideHeaders;
+
+export const initializeServerSideHeaders = (headers) => {
+  serverSideHeaders = headers || {};
+};
+
+export const getServerSideHeader = (headerName) => {
+  if (!serverSideHeaders) {
+    return '';
+  }
+
+  return serverSideHeaders[headerName] || '';
+};
